requirements.txt
Django==4.2.7
django-ninja==1.0.1
djangorestframework==3.14.0
python-decouple==3.8
asgiref==3.7.2
uvicorn==0.24.0